Extract src/palui.ts — the ⌘K palette's UI
First of evidence group 2 (cosmetic: these render on demand, never from
renderAll, so isolating them tells the Phase-3 profiling pass nothing).

The split mirrors tasks.ts / taskui.ts: ./palette already holds this
surface's decisions under 39 tests — fuzzy match, scoring, prefix parsing,
frecency — so what moved here is only the box, the group assembly, the row
markup and the key handling. That is why the slice adds no tests.

Moved whole (two guarded ranges): PalGroup, the five pieces of palette
state, sessionActions, PAL_CMDS, buildPalGroups, runPalItem, openPalActions,
popPalPage, renderPal, refreshPal, palTasks, openPalette, closePalette, and
the two $("palInput") listeners — which address rows by index into palFlat,
so per the DOM-owning recipe they were never part of the global dispatcher
and belong here. The global ⌘K toggle stays in main.ts.

**One host object, not per-callee setters.** A palette row can do almost
anything the app can do: twelve of its callees live in main.ts (setActive,
resolvePermission, runGit, openPlainTerminal, closeSession, addProject,
cycleSort, toggleInsp, toggleRail, toggleTheme, requestLaunch,
openInputPrompt). PLAN allows this for a control panel and settings.ts is
the precedent; twelve setters would be noise. Stated here, as PLAN asks.

The rewrite to `host.` was scripted with a guard that refuses to run if the
name is declared in the file — PLAN's recorded trap is a regex call-rewrite
that also renames a declaration — and it had to cover bare references, not
just calls: a PAL_CMDS row stores `run: addProject`. Every one of the twenty
sites it touched is in the script's output and in the verification diff.

Verified mechanical: the relocated block diffed against
`git show HEAD:src/main.ts` differs only by those twenty `host.` prefixes,
two `export` keywords, and a deleted section divider that the new module
docblock now says. main.ts 2,074 -> 1,910. Gates: tsc --noEmit clean, 368
vitest pass. No Rust touched.
